Migrate ExamListForm to DoctorExamsForm
